+++ backend/.env.sample
@@ -21,6 +21,4 @@ AIOD_KEYCLOAK__REALM=dev
 AIOD_KEYCLOAK__CLIENT_ID=<YOUR_CLIENT_ID>
 AIOD_KEYCLOAK__CLIENT_SECRET=<YOUR_CLIENT_SECRET>
 AIOD_KEYCLOAK__SERVER_URL=https://aiod-dev.i3a.es/aiod-auth/
-AIOD_KEYCLOAK__OIDC_URL=https://aiod-dev.i3a.es/aiod-auth/realms/dev/.well-known/openid-configuration
-
-PASSWORD_FOR_TEMPLATE_APPROVAL="pass"
+AIOD_KEYCLOAK__OIDC_URL=https://aiod-dev.i3a.es/aiod-auth/realms/aiod/.well-known/openid-configuration
