+++ backend/app/config.py
@@ -38,8 +38,6 @@ class AIODKeycloakConfig(BaseModel):
 
 
 class Settings(BaseSettings):
-    PASSWORD_FOR_TEMPLATE_APPROVAL: str
-
     MONGODB_URI: str
     MONGODB_DBNAME: str
