+++ backend/app/main.py
@@ -12,7 +12,7 @@ from app.config import TEMP_DIRNAME, settings
 from app.models.experiment import Experiment
 from app.models.experiment_run import ExperimentRun
 from app.models.experiment_template import ExperimentTemplate
-from app.routers import aiod, experiment_runs, experiment_templates, experiments
+from app.routers import admin, aiod, experiment_runs, experiment_templates, experiments
 from app.services.aiod import aiod_client_wrapper, aiod_library_client_wrapper
 from app.services.container_platforms.base import ContainerPlatformBase
 from app.services.container_platforms.docker import DockerService
@@ -28,6 +28,7 @@ app.include_router(
 )
 app.include_router(experiments.router, prefix="/v1", tags=["experiments"])
 app.include_router(experiment_runs.router, prefix="/v1", tags=["experiment-runs"])
+app.include_router(admin.router, prefix="/v1", tags=["admin"])
 
 app.add_middleware(
     CORSMiddleware,
