+++ backend/pytest.ini
@@ -28,5 +28,3 @@ env =
     AIOD_KEYCLOAK__CLIENT_SECRET=
     AIOD_KEYCLOAK__SERVER_URL=https://keycloak.aiod.eu
     AIOD_KEYCLOAK__OIDC_URL=https://keycloak.aiod.eu/realms/aiod/.well-known/openid-configuration
-
-    PASSWORD_FOR_TEMPLATE_APPROVAL=
